+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
+- `sandstone/ContextualPopupDecorator` to update popup position properly when Wrapped component updated
 - `sandstone/Steps` to update opacity of future steps to match the GUI guide
 
 ## [2.9.10] - 2025-02-24